resume_analyzer/requirements.txt
flask==3.0.3
PyPDF2==3.0.1
python-docx==1.1.2
spacy==3.7.4
nltk==3.8.1
scikit-learn==1.4.2
pandas==2.2.2
numpy==1.26.4
werkzeug==3.0.3
